# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing data on oncological markers.",
        "title": "OncomarkerAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/procedure-blood-count": {
            "get": {
                "description": "Retrieves a list of procedure blood count entries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Get procedure blood count list",
                "responses": {
                    "200": {
                        "description": "Procedure blood count list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.ProcedureBloodCount"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing procedure blood count entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Update procedure blood count",
                "parameters": [
                    {
                        "description": "Procedure blood count data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProcedureBloodCount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated procedure blood count data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new procedure blood count entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Create procedure blood count",
                "parameters": [
                    {
                        "description": "Procedure blood count data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProcedureBloodCount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created procedure blood count data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/procedure-blood-count/blood-counts/{blood_count_id}": {
            "get": {
                "description": "Retrieves a list of procedure blood count entries by blood count ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Get procedure blood count list by blood count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Procedure blood count list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.ProcedureBloodCount"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/procedure-blood-count/procedures/{procedure_id}": {
            "get": {
                "description": "Retrieves a list of procedure blood count entries by procedure ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Get procedure blood count list by procedure",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Procedure ID",
                        "name": "procedure_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Procedure blood count list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.ProcedureBloodCount"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/procedure-blood-count/procedures/{procedure_id}/blood-counts/{blood_count_id}": {
            "get": {
                "description": "Retrieves a procedure blood count entry by procedure ID and blood count ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Get procedure blood count by IDs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Procedure ID",
                        "name": "procedure_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Procedure blood count data",
                        "schema": {
                            "$ref": "#/definitions/model.ProcedureBloodCount"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a procedure blood count entry by procedure ID and blood count ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProcedureBloodCount"
                ],
                "summary": "Delete procedure blood count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Procedure ID",
                        "name": "procedure_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blood count ID",
                        "name": "blood_count_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted procedure blood count data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ProcedureBloodCount": {
            "type": "object",
            "properties": {
                "blood-count": {
                    "type": "string"
                },
                "measure-code": {
                    "type": "string"
                },
                "procedure": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}